[Update] 🔥 상품평 등록 후 화면 reload

+++ GoGuMa/src/main/java/com/ggm/goguma/service/product/ReviewService.java
@@ -15,5 +15,5 @@ public interface ReviewService {
 	void insertReview(long productID, long memberID, String content) throws Exception;
 	
 	// 상품평 등록 자격 확인
-	long isFinishRcpt(long productID, long memberID) throws Exception;
+	Integer isFinishRcpt(long productID, long memberID) throws Exception;
 }
